@@ -2,6 +2,7 @@ import { Injectable, CanActivate, ExecutionContext } from '@nestjs/common';
 import { Observable } from 'rxjs';
 import { AuthGuard } from './auth.guard';
 import * as jwt from 'jsonwebtoken';
+import { GqlExecutionContext } from '@nestjs/graphql';
 const secret  = 'GRAPHQL_CONFERENCE_SECRET';
 
 @Injectable()
@@ -9,8 +10,9 @@ export class RoleGuard extends AuthGuard {
   canActivate(
     context: ExecutionContext,
   ): boolean | Promise<boolean> | Observable<boolean> {
-    const request = context.switchToHttp().getRequest();
-    const authorization = request.headers['authorization'];
+    const ctx = GqlExecutionContext.create(context);
+    const { req } = ctx.getContext();
+    const authorization = req.headers['authorization'];
     return authorization != null && this.hasRole(authorization, 'superpower');
   }
 
